Replace pycodestyle with flake8 (#19)
Previously, we would run *pycodestyle* through *pytest*. This is not
ideal because it connects static analysis with actual testing. Additionally,
previous runs showed that *pycodestyle* would not discover violations
like unused imports or "bad" non-PEP8 names.

Now, the linting is mostly unified with the library version (except that we
are still missing strict typing for the CLI). I decided to raise the line length
from 79 to 160 characters in the same step to improve readability, especially
since most devices are large enough nowadays to still display this in one line
without soft wrapping.